requirements.txt
fastapi
uvicorn[standard]
concrete-python
pydantic
